+++ package.json
@@ -18,7 +18,7 @@
     "js-cookie": "^3.0.1",
     "katex": "^0.13.18",
     "react": "^16.8.5",
-    "react-dom": "16.8.4",
+    "react-dom": "16.8.5",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
