+++ src/components/semantic/BaseValuePresenter.tsx
@@ -144,6 +144,7 @@ export const ValueWrapper: FunctionComponent<ValueWrapperProps> = ({
         if (e.target === e.currentTarget) {
             valueRef.current?.startEdit();
         }
+        console.log("ValueWrapper click", e);
     })}>
         {children}
     </div>;
