+++ src/components/semantic/audience.module.css
@@ -1,0 +1,3 @@
+.wrapper button {
+    margin-left: 0.5em;
+}
